Added anchors and began refactoring Geometry
Anchors are reactive structures with a position and a velocity. They
allow one entity in the world to be fixed either to an absolute position
in the world, an absolute position that changes over time, to another
entity in the world, or some composition of any of those. Anchors are
only reactive in one direction, however, so anchoring one entity to
another means the anchored entity cannot move independent of its anchor.

The refactoring of Geometry isn't complete yet, so I'll save the
description of that for a later commit.