Changed Geometry, continued working on splines.
After a lot of flip-flopping, I decided that having a family of Quad
classes was more than necessary, and that a Rectangle class should be
enough. Started working on creating bounding volumes for splines so that
rectangle-spline intersection can be calculated, but I'll save a full
description for a later commit.